+++ hutool-core/src/test/java/cn/hutool/core/exceptions/ExceptionUtilTest.java
@@ -27,6 +27,10 @@ public class ExceptionUtilTest {
 	}
 
 	@Test
+	public void getStackElementTest(){
+	}
+
+	@Test
 	public void convertTest() {
 		// RuntimeException e = new RuntimeException();
 		IOException ioException = new IOException();