+++ hutool-log/src/main/java/cn/hutool/log/dialect/logtube/LogTubeLogFactory.java
@@ -1,0 +1,29 @@
+package cn.hutool.log.dialect.logtube;
+
+import cn.hutool.log.Log;
+import cn.hutool.log.LogFactory;
+
+/**
+ * <a href="https://github.com/logtube/logtube-java">LogTube</a> log. 封装<br>
+ *
+ * @author Looly
+ * @since 5.6.6
+ */
+public class LogTubeLogFactory extends LogFactory {
+
+	public LogTubeLogFactory() {
+		super("LogTube");
+		checkLogExist(io.github.logtube.Logtube.class);
+	}
+
+	@Override
+	public Log createLog(String name) {
+		return new LogTubeLog(name);
+	}
+
+	@Override
+	public Log createLog(Class<?> clazz) {
+		return new LogTubeLog(clazz);
+	}
+
+}